backend/requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
flask-cors==4.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
PyJWT==2.8.0
Werkzeug==3.1.5
psycopg2-binary==2.9.11
pytest==9.0.2
requests==2.32.5
